name app and old app correctly, add redux + basic UI
