+++ src/components/App/actionTypes.js
@@ -1,2 +1,1 @@
-export const SET_USERS = 'SET_USERS';
-export const SET_SLACK_USER_FETCH_ERROR = 'SET_SLACK_USER_FETCH_ERROR';
+export const SET_SEND_WITH_PHOTO = 'SET_SEND_WITH_PHOTO';
