+++ src/components/App/actions.js
@@ -1,22 +1,8 @@
-import {SET_USERS, SET_SLACK_USER_FETCH_ERROR} from './actionTypes';
+import {SET_SEND_WITH_PHOTO} from './actionTypes';
 
-import {loadUsers as reallyLoadUsers} from './../../utils/slack';
-
-export function setUsers(users) {
+export function setSendWithPhoto(sendWithPhoto) {
 	return {
-		type: SET_USERS,
-		users
+		type: SET_SEND_WITH_PHOTO,
+		sendWithPhoto
 	};
 }
-
-export function setSlackUserFetchError(slackUserFetchError) {
-	return {
-		type: SET_SLACK_USER_FETCH_ERROR,
-		slackUserFetchError
-	};
-}
-
-export const loadUsers = () => (dispatch, getState) =>
-	reallyLoadUsers()
-		.then(users => dispatch(setUsers(users)))
-		.catch(() => dispatch(setSlackUserFetchError(true)));
