+++ src/components/App/reducer.js
@@ -1,12 +1,6 @@
-import {SET_USERS, SET_SLACK_USER_FETCH_ERROR} from './actionTypes';
+import {SET_SEND_WITH_PHOTO} from './actionTypes';
 
-export function users(state = [], action) {
-	if (action.type === SET_USERS) return action.users;
-	return state;
-}
-
-export function slackUserFetchError(state = false, action) {
-	if (action.type === SET_SLACK_USER_FETCH_ERROR)
-		return action.slackUserFetchError;
+export function sendWithPhoto(state = true, action) {
+	if (action.type === SET_SEND_WITH_PHOTO) return action.sendWithPhoto;
 	return state;
 }
