--- src/components/NewApp/NewApp.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const NewApp = props => {
-	return <div />;
-};
-
-export default NewApp;
