+++ src/components/OldApp/actionTypes.js
@@ -1,0 +1,2 @@
+export const SET_USERS = 'SET_USERS';
+export const SET_SLACK_USER_FETCH_ERROR = 'SET_SLACK_USER_FETCH_ERROR';
