+++ src/index.js
@@ -2,20 +2,20 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import './index.css';
-import AppContainer from './components/App/AppContainer';
+import OldAppContainer from './components/OldApp/OldAppContainer';
 import store from './utils/reduxStore';
 import registerServiceWorker from './registerServiceWorker';
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
-import NewApp from './components/NewApp/NewApp';
+import AppContainer from './components/App/container';
 import Trainer from './components/Trainer/Trainer';
 
 ReactDOM.render(
 	<Provider store={store}>
 		<Router>
 			<Switch>
-				<Route path="/" component={NewApp} />
-				<Route path="/old" component={AppContainer} />
-				<Route path="/training" component={Trainer} />
+				<Route exact path="/" component={AppContainer} />
+				<Route exact path="/old" component={OldAppContainer} />
+				<Route exact path="/training" component={Trainer} />
 			</Switch>
 		</Router>
 	</Provider>,
